pyproject.toml
[project]
name = "invoice-graph-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=1.2.9",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.8",
    "pymupdf>=1.26.7",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
]
